Add modType name/value pair to spec
This separates the `type` of module (i.e., content or profile) from a more specific indication of what is contained in the module. 
For a content module, one might indicate 'dataset' as a more specific indication. Per our recommendation, the Wikidata identifier 'Q1172284' is used (this helps prevent the need to disambiguate 'dataset' and 'data set', 'data-set', etc.)